pyproject.toml
[project]
name = "D4G"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camelot-py>=0.11.0",
    "ghostscript>=0.7",
    "ipykernel>=6.29.5",
    "jpype1>=1.5.0",
    "llama-index-core>=0.12.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-readers-file>=0.4.0",
    "llama-parse>=0.5.14",
    "matplotlib>=3.9.2",
    "nbformat>=5.10.4",
    "nest-asyncio>=1.6.0",
    "numpy>=2.1.3",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.40.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tabula-py>=2.10.0",
]
